explicit export on public-api to remove warnings
