+++ projects/lux/src/public-api.ts
@@ -2,15 +2,27 @@
  * Public API Surface of lux
  */
 
-export * from './lib/lux.module';
-export * from './lib/autocomplete/autocomplete.component';
-export * from './lib/autocomplete-list/autocomplete-list.component';
-export * from './lib/breadcrumb/breadcrumb.component';
-export * from './lib/checkbox/checkbox.component';
-export * from './lib/filter/filter.component';
-export * from './lib/input/input.component';
-export * from './lib/pagination/pagination.component';
-export * from './lib/tooltip/tooltip.directive';
-export * from './lib/modal/modal.service';
-export * from './lib/radiogroup/radiogroup.component';
+export { LuxModule } from './lib/lux.module';
+export {
+  DataSource,
+  DataSourceItem,
+  DecoratedDataSource,
+  DecoratedDataSourceItem
+} from './lib/datasource';
+export { AutocompleteComponent } from './lib/autocomplete/autocomplete.component';
+export { AutocompleteListComponent } from './lib/autocomplete-list/autocomplete-list.component';
+export {
+  LuxBreadcrumbComponent,
+  BreadcrumbItem
+} from './lib/breadcrumb/breadcrumb.component';
+export { CheckboxComponent } from './lib/checkbox/checkbox.component';
+export { FilterComponent } from './lib/filter/filter.component';
+export { InputComponent } from './lib/input/input.component';
+export { PaginationComponent } from './lib/pagination/pagination.component';
+export { LuxTooltipDirective } from './lib/tooltip/tooltip.directive';
+export { ModalService } from './lib/modal/modal.service';
+export {
+  RadioItem,
+  RadiogroupComponent
+} from './lib/radiogroup/radiogroup.component';
 export { ModalRef } from './lib/modal/modal-ref';
